first adjustments, extraction is buggy, probably smth with the discriminant
