+++ Cargo.toml
@@ -16,10 +16,11 @@ name = "eggshell"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
+bon = "2"
 chrono = "0.4"
 clap = { version = "4", features = ["derive"] }
 csv = "1"
-egg = { version = "0.9", features = ["serde-1"] }
+egg = { git = "https://github.com/SaltyPeppermint/egg.git" , features = ["serde-1"] }
 env_logger = "0.11.5"
 hashbrown = { version = "0.14", features = ["rayon", "serde"] }
 indexmap = { version = "2", features = ["serde"] }
