+++ src/sketch/utils.rs
@@ -100,15 +100,15 @@ where
 
 /// Workaround since an identical data struct `classes_by_op` is private in the egraph struct
 /// Using just the discriminant is ok since it is again checked in the `for_each_matching_node` function
-pub fn new_classes_by_op<L, N>(egraph: &EGraph<L, N>) -> HashMap<Discriminant<L>, HashSet<Id>>
+pub fn new_classes_by_op<L, N>(egraph: &EGraph<L, N>) -> HashMap<L::Discriminant, HashSet<Id>>
 where
     L: Language,
     N: Analysis<L>,
 {
-    let mut classes_by_op: HashMap<Discriminant<L>, HashSet<Id>> = HashMap::default();
+    let mut classes_by_op: HashMap<L::Discriminant, HashSet<Id>> = HashMap::default();
     for class in egraph.classes() {
         for node in &class.nodes {
-            let key = std::mem::discriminant(node);
+            let key = node.discriminant();
             classes_by_op
                 .entry(key)
                 .and_modify(|ids| {
@@ -121,9 +121,9 @@ where
 }
 
 pub fn classes_matching_op<'a, L: Language>(
-    enode: &'a L,
-    classes_by_op: &'a HashMap<Discriminant<L>, HashSet<Id>>,
+    enode: &L,
+    classes_by_op: &'a HashMap<L::Discriminant, HashSet<Id>>,
 ) -> Option<&'a HashSet<Id>> {
-    let key = std::mem::discriminant(enode);
+    let key = enode.discriminant();
     classes_by_op.get(&key)
 }
